requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
colorama==0.4.6
    # via crayons
crayons==0.4.0
    # via stegano
exitstatus==2.5.0
    # via -r requirements.in
numpy==2.0.0
    # via opencv-python
opencv-python==4.10.0.84
    # via stegano
piexif==1.1.3
    # via stegano
pillow==10.3.0
    # via stegano
pycryptodomex==3.20.0
    # via -r requirements.in
stegano==0.11.3
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
